TS: Table/HighlihtToggler
