+++ src/components/Table/HighlightToggler/style.css
@@ -1,17 +1,20 @@
-@import '../../../styles/colors.css';
+@import "../../../styles/colors.css";
 
-.highlight-toggler {
+body:global(.interpro) .highlight-toggler {
   padding: 0.1em;
   cursor: pointer;
-}
-.highlight-toggler:focus {
-  border: 0;
-  outline-width: 0;
-}
-.highlight-toggler::before {
-  background: var(--colors-graydark);
-  color: var(--colors-nearly-white);
-}
-.highlight-toggler.on::before {
-  color: yellow;
+  &:focus {
+    border: 0;
+    outline-width: 0;
+  }
+  &::before {
+    background-color: var(--colors-graydark);
+    color: var(--colors-nearly-white);
+  }
+  &.on::before {
+    color: yellow;
+  }
+  &:hover::before {
+    background-color: var(--colors-gray);
+  }
 }
